Use python3 in the system environment
Some user setups are expecting to use a python in a different location
than `/usr/bin`, and an older python interpreter may be present at that
location (Rocky Linux 8 being one), causing issues (notably that
shutil.copytree does not have `dirs_exist_ok` keyword parameter in
Python versions < 3.8).

This uses the `python3` interpreter that is in the shell's current PATH,
as discovered by `meson` instead of hard-coding it to
`/usr/bin/python3`.

https://gitlab.gnome.org/GNOME/gnome-browser-extension/-/issues/50